scripts/requirements.txt
# Data processing
numpy>=1.26.0
requests>=2.31.0

# Coordinate transformation
pyproj>=3.6.0

# 3D mesh processing
trimesh>=4.0.0
scipy>=1.11.0

# Raster/terrain processing
rasterio>=1.3.0
pillow>=10.0.0

# GeoJSON handling
shapely>=2.0.0
geojson>=3.0.0

# XML parsing (for WC Guide toilets data)
lxml>=5.0.0

# GPKG/GeoPackage processing
geopandas>=0.14.0
fiona>=1.9.0

# Progress bars
tqdm>=4.66.0

# Tile pipeline (optional but recommended)
rtree>=1.0.0        # Spatial indexing for vector queries
pysolar>=0.10       # Sun position calculations
noise>=1.2.2        # Perlin noise for hand-drawn wobble effect

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# AI-powered tile generation (optional)
# Uses existing deps: requests, numpy, pillow, scipy
# Requires API keys:
#   - GOOGLE_API_KEY for Gemini (https://aistudio.google.com)
#   - OPENAI_API_KEY for GPT Image (https://platform.openai.com)
#   - REPLICATE_API_TOKEN for Stable Diffusion (https://replicate.com)

# LiteLLM - multi-provider AI abstraction
litellm>=1.76.0     # Unified API for Gemini, OpenAI, Vertex AI image generation

# Stable Diffusion via Replicate (for ControlNet Depth pipeline)
replicate>=0.25.0

# Shadow removal + Ray tracing pipeline (optional but recommended)
simple-lama-inpainting>=0.1.0  # AI inpainting for shadow removal
mapbox-earcut>=1.0.0           # Polygon triangulation for 3D scene building
# pyembree>=0.1.0              # Optional: faster ray tracing (Intel CPUs only)
